refactor: extract link-detection helpers into ui/links.rs (#357)
Sixteenth slice. Final extraction of #357: pulls the link/URL
detection helpers out of ui/mod.rs into a dedicated submodule.

Moved:
- LinkRegion struct          (re-exported as pub via ui/mod.rs)
- collect_link_regions       (used by ui::draw post-render OSC 8 emit)
- is_link_style              (private)
- extract_url                (private)
- split_spans_by_newline     (used by chat_pane via super::links)
- styled_uri_spans           (used by chat_pane via super::links)
- extract_url_cases tests    (now in links.rs's tests module)

ui/mod.rs is now 757 lines, down from the original 5,272 (an 86%
reduction across 16 PRs). The final shape of the ui module:
  ui/
    mod.rs              top-level draw() + shared helpers
    autocomplete.rs     popup above the composer
    chat_pane.rs        messages list + image patch pass
    composer.rs         input box
    links.rs            link detection + URI styling
    sidebar.rs          conversation list
    status_bar.rs       bottom status line
    welcome.rs          empty-state screen
    overlays/           19 modal overlay submodules

Closes #357.
